Fix: Move exceptions into Exception namespace
